Added temp config system.
