--- arduino/arduino_controller/arduino_controller.ino
@@ -1,9 +1,0 @@
-void setup() {
-  // put your setup code here, to run once:
-
-}
-
-void loop() {
-  // put your main code here, to run repeatedly:
-
-}
