--- arduino/config.json
@@ -1,14 +1,0 @@
-{
-	"LED_PIN": 5,
-	"NUM_LEDS": 63,
-	"BRIGHTNESS": 64,
-	"LED_TYPE": "WS2811",
-
-	"COLOR_ORDER": "GRB",
-	"UPDATES_PER_SECOND": 100,
-
-	"PATTERNS": [
-		"blink"
-	]
-
-}
